Overwrite cached predictions on demand

+++ utils.py
@@ -1,3 +1,5 @@
+import os
+
 from datetime import datetime
 
 import numpy as np
@@ -5,10 +7,10 @@ import numpy as np
 from src.dataset import Traffic4CastSample
 
 
-def cache(func, path, *args, **kwargs):
-    try:
+def cache(func, path, to_overwrite, *args, **kwargs):
+    if os.path.exists(path) and not to_overwrite:
         return np.load(path)
-    except FileNotFoundError:
+    else:
         result = func(*args, **kwargs)
         np.save(path, result)
         return result
